Move markets and reminder packages under news as subpackages
Markets and reminder are conceptually subsets of the news feature.
Move them as news/markets and news/reminder, updating imports in
main.go, home/home.go, and blog/digest/digest.go.
