+++ blog/digest/digest.go
@@ -10,9 +10,9 @@ import (
 	"mu/app"
 	"mu/blog"
 	"mu/data"
-	"mu/markets"
 	"mu/news"
-	"mu/reminder"
+	"mu/news/markets"
+	"mu/news/reminder"
 	"mu/video"
 )
