+++ main.go
@@ -24,10 +24,10 @@ import (
 	"mu/docs"
 	"mu/home"
 	"mu/mail"
-	"mu/markets"
 	"mu/news"
+	"mu/news/markets"
+	"mu/news/reminder"
 	"mu/places"
-	"mu/reminder"
 	"mu/search"
 	"mu/user"
 	"mu/video"
